Install a specific branch or tag with `@` syntax
- Use idiomatic `username/package@ref` style syntax.
- Check for upstream changes in a more portable way
- Disable outdated check on packages cloned by ref or tag

The ref selection is permanent at install time. Theres no way to
switch the package to a different ref without re-installing it.

Tags in git are effectively non-updatable. If you checked out a package
on a tag or commit hash its very likely that you'll never expect that
tag to change and want to pin the package to that specific version

Thanks to osse from #git for quick help :heart: