Fix the find command in link-bins
Under the following conditions, basher-_link-bins would think
that the package itself is a bin, to be symlinked and chmod'ed,
and would fail.

The conditions are:
- the package was installed with `basher link`
- the package did not specify BINS in package.sh
- the package had no bin directory

This fix simply adds the -mindepth 1 option to the find command
responsible for the issue. It tells find to never consider the
directory/package itself as a matching element.