feat: Support [un]installation of multiple packages
When multiple packages are specified on the command line, for the 'install'
and 'uninstall' subcommands, it iterates through each one in order, installing
them one after the next. If one package fails to install, installation is halted
and the usual error shows on screen. This also the tests checking for extra
arguments, since those are properly handled now